FS-9951: [mod_fifo] Add early media support to fifo in
With this changes there is a 4th optional parameter in application
'fifo in' which can be set to 'early' or 'noans'.

With noans, mod_fifo never answer on a-leg channel : this must be done
in dialplan.

With early, mod_fifo answer on a-leg only when b-leg has been answered : this allow
a-leg to stay in early when it is put in fifo .