FS-10618: [mod_sofia] Don't cancel outbound channels on nightmare transfer during originate
When performing nightmare transfer during originate and if multiple endpoints
are called, only the last one will be intercepted and will continue to ring,
while the rest of the outbound channels will be cancelled. When performing a
non-nightmare transfer, all of the called endpoints continue to ring, which is
the expected behaviour.

This patch fixes the nightmare transfer case, so it works the same as the
non-nightmare transfer case.

FS-10618 #resolve